[#0] moved breadcrumb logic back to block for SEO
